(tests): Move record parser and its tests to their own files
