+++ txtdirect.go
@@ -37,18 +37,6 @@ const (
 	status301CacheAge = 604800
 )
 
-type record struct {
-	Version string
-	To      string
-	Code    int
-	Type    string
-	Vcs     string
-	Website string
-	From    string
-	Root    string
-	Re      string
-}
-
 // Config contains the middleware's configuration
 type Config struct {
 	Enable     []string
@@ -57,96 +45,6 @@ type Config struct {
 	LogOutput  string
 	Gomods     Gomods
 	Prometheus Prometheus
-}
-
-// Parse takes a string containing the DNS TXT record and returns
-// a TXTDirect record struct instance.
-// It will return an error if the DNS TXT record is not standard or
-// if the record type is not enabled in the TXTDirect's config.
-func (r *record) Parse(str string, req *http.Request, c Config) error {
-	s := strings.Split(str, ";")
-	for _, l := range s {
-		switch {
-		case strings.HasPrefix(l, "code="):
-			l = strings.TrimPrefix(l, "code=")
-			i, err := strconv.Atoi(l)
-			if err != nil {
-				return fmt.Errorf("could not parse status code: %s", err)
-			}
-			r.Code = i
-
-		case strings.HasPrefix(l, "from="):
-			l = strings.TrimPrefix(l, "from=")
-			l, err := parsePlaceholders(l, req, []string{})
-			if err != nil {
-				return err
-			}
-			r.From = l
-
-		case strings.HasPrefix(l, "re="):
-			l = strings.TrimPrefix(l, "re=")
-			r.Re = l
-
-		case strings.HasPrefix(l, "root="):
-			l = strings.TrimPrefix(l, "root=")
-			r.Root = l
-
-		case strings.HasPrefix(l, "to="):
-			l = strings.TrimPrefix(l, "to=")
-			l, err := parsePlaceholders(l, req, []string{})
-			if err != nil {
-				return err
-			}
-			r.To = l
-
-		case strings.HasPrefix(l, "type="):
-			l = strings.TrimPrefix(l, "type=")
-			r.Type = l
-
-		case strings.HasPrefix(l, "v="):
-			l = strings.TrimPrefix(l, "v=")
-			r.Version = l
-			if r.Version != "txtv0" {
-				return fmt.Errorf("unhandled version '%s'", r.Version)
-			}
-			log.Print("WARN: txtv0 is not suitable for production")
-
-		case strings.HasPrefix(l, "vcs="):
-			l = strings.TrimPrefix(l, "vcs=")
-			r.Vcs = l
-
-		case strings.HasPrefix(l, "website="):
-			l = strings.TrimPrefix(l, "website=")
-			r.Website = l
-
-		default:
-			tuple := strings.Split(l, "=")
-			if len(tuple) != 2 {
-				return fmt.Errorf("arbitrary data not allowed")
-			}
-			continue
-		}
-		if len(l) > 255 {
-			return fmt.Errorf("TXT record cannot exceed the maximum of 255 characters")
-		}
-		if r.Type == "dockerv2" && r.To == "" {
-			return fmt.Errorf("[txtdirect]: to= field is required in dockerv2 type")
-		}
-	}
-
-	if r.Code == 0 {
-		r.Code = http.StatusFound
-	}
-
-	if r.Type == "" {
-		r.Type = "host"
-	}
-
-	if !contains(c.Enable, r.Type) {
-		return fmt.Errorf("%s type is not enabled in configuration", r.Type)
-	}
-
-	return nil
 }
 
 // getBaseTarget parses the placeholder in the given record's To= field
@@ -171,39 +69,6 @@ func contains(array []string, word string) bool {
 		}
 	}
 	return false
-}
-
-// getRecord uses the given host to find a TXT record
-// and then parses the txt record and returns a TXTDirect record
-// struct instance. It returns an error when it can't find any txt
-// records or if the TXT record is not standard.
-func getRecord(host string, ctx context.Context, c Config, r *http.Request) (record, error) {
-	txts, err := query(host, ctx, c)
-	if err != nil {
-		log.Printf("Initial DNS query failed: %s", err)
-	}
-	// if error present or record empty, jump into wildcards
-	if err != nil || txts[0] == "" {
-		hostSlice := strings.Split(host, ".")
-		hostSlice[0] = "_"
-		host = strings.Join(hostSlice, ".")
-		txts, err = query(host, ctx, c)
-		if err != nil {
-			log.Printf("Wildcard DNS query failed: %s", err.Error())
-			return record{}, err
-		}
-	}
-
-	if len(txts) != 1 {
-		return record{}, fmt.Errorf("could not parse TXT record with %d records", len(txts))
-	}
-
-	rec := record{}
-	if err = rec.Parse(txts[0], r, c); err != nil {
-		return rec, fmt.Errorf("could not parse record: %s", err)
-	}
-
-	return rec, nil
 }
 
 // fallback redirects the request to the given fallback address
